epoqueart/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-admin-interface==0.18.5
django-appconf==1.0.5
django-colorfield==0.6.3
django-compressor
django-compressor-toolkit
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-embed-video==1.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-routes==0.2.0
django-password-validators==1.6.1
django-session-timeout==0.1.0
django-libsass
django-compressor-autoprefixer
phonenumbers
django-phonenumber-field
gunicorn==20.1.0
idna==3.3
Pillow
ply==3.11
psycopg2==2.9.3
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.8
zope.event==4.5.0
zope.interface==5.4.0